# xmcy/controller/src/main/resources/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Restful",
    "version": "1.0",
    "title": "Spring Boot with Swagger2 to create RESTful API",
    "termsOfService": "http://epam.com",
    "contact": {
      "name": "Ryhor Pishchyk",
      "url": "http://epam.com",
      "email": "ryhor_pishchyk@epam.com"
    }
  },
  "host": "localhost:8080",
  "basePath": "/",
  "tags": [
    {
      "name": "Module for uploading cryptos rate find information by request",
      "description": "Crypto Controller"
    },
    {
      "name": "Module to find normalized information by request",
      "description": "Normalized Crypto Controller"
    }
  ],
  "paths": {
    "/crypto/maxByMonth": {
      "get": {
        "tags": [
          "Module for uploading cryptos rate find information by request"
        ],
        "summary": "Find max rates by month",
        "operationId": "findMaxByMonthUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "month",
            "description": "month",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ResponseEntity"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/crypto/maxByMonthAndSymbol": {
      "get": {
        "tags": [
          "Module for uploading cryptos rate find information by request"
        ],
        "summary": "Find max rates by month and name of the cryptocurrency",
        "operationId": "findMaxByMonthAndSymbolUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "month",
            "description": "month",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "body",
            "name": "symbol",
            "description": "symbol",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ResponseEntity"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/crypto/mintByMonth": {
      "get": {
        "tags": [
          "Module for uploading cryptos rate find information by request"
        ],
        "summary": "Find min rates by month",
        "operationId": "findMinByMonthUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "month",
            "description": "month",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ResponseEntity"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/crypto/mintByMonthAndSymbol": {
      "get": {
        "tags": [
          "Module for uploading cryptos rate find information by request"
        ],
        "summary": "Find min rates by month and name of the cryptocurrency",
        "operationId": "findMinByMonthAndSymbolUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "month",
            "description": "month",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "body",
            "name": "symbol",
            "description": "symbol",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ResponseEntity"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/crypto/newestByMonth": {
      "get": {
        "tags": [
          "Module for uploading cryptos rate find information by request"
        ],
        "summary": "Find most newest rates by month",
        "operationId": "findNewestByMonthUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "month",
            "description": "month",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ResponseEntity"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/crypto/newestByMonthAndSymbol": {
      "get": {
        "tags": [
          "Module for uploading cryptos rate find information by request"
        ],
        "summary": "Find most newest rates by month and name of the cryptocurrency",
        "operationId": "findNewestByMonthAndSymbolUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "month",
            "description": "month",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "body",
            "name": "symbol",
            "description": "symbol",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ResponseEntity"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/crypto/oldestByMonth": {
      "get": {
        "tags": [
          "Module for uploading cryptos rate find information by request"
        ],
        "summary": "Find most oldest rates by month",
        "operationId": "findOldestByMonthUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "month",
            "description": "month",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ResponseEntity"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/crypto/oldestByMonthAndSymbol": {
      "get": {
        "tags": [
          "Module for uploading cryptos rate find information by request"
        ],
        "summary": "Find most oldest rates by month and name of the cryptocurrency",
        "operationId": "findOldestByMonthAndSymbolUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "month",
            "description": "month",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "body",
            "name": "symbol",
            "description": "symbol",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ResponseEntity"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/crypto/upload": {
      "post": {
        "tags": [
          "Module for uploading cryptos rate find information by request"
        ],
        "summary": "Upload SCV file with cryptos rate",
        "operationId": "uploadSalaryScvFileUsingPOST",
        "consumes": [
          "multipart/form-data"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "file",
            "description": "file",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ResponseEntity"
            }
          },
          "201": {
            "description": "Created"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/normalized/byMonth": {
      "get": {
        "tags": [
          "Module to find normalized information by request"
        ],
        "summary": "Find a descending sorted list of all the cryptos, comparing the normalized range",
        "operationId": "findNormalizedByMonthUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "month",
            "description": "month",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ResponseEntity"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/normalized/byMonthAndDay": {
      "get": {
        "tags": [
          "Module to find normalized information by request"
        ],
        "summary": "Find crypto with the highest normalized range for a specific day",
        "operationId": "findNormalizedByMonthAndDayUsingGET",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "*/*"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "month",
            "description": "month",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "body",
            "name": "day",
            "description": "day",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/ResponseEntity"
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    }
  },
  "definitions": {
    "CryptoDto": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "price": {
          "type": "number"
        },
        "symbol": {
          "type": "string"
        },
        "timestamp": {
          "type": "integer",
          "format": "int64"
        }
      }
    },
    "NormalizedCryptoDto": {
      "type": "object",
      "properties": {
        "max": {
          "type": "number"
        },
        "min": {
          "type": "number"
        },
        "normalizedRate": {
          "type": "number"
        },
        "symbol": {
          "type": "string",
          "enum": [
            "BTC",
            "DOGE",
            "ETH",
            "LTC",
            "XRP"
          ]
        }
      }
    },
    "ResponseEntity": {
      "type": "object",
      "properties": {
        "body": {
          "type": "object"
        },
        "statusCode": {
          "type": "string",
          "enum": [
            "100",
            "101",
            "102",
            "103",
            "200",
            "201",
            "202",
            "203",
            "204",
            "205",
            "206",
            "207",
            "208",
            "226",
            "300",
            "301",
            "302",
            "303",
            "304",
            "305",
            "307",
            "308",
            "400",
            "401",
            "402",
            "403",
            "404",
            "405",
            "406",
            "407",
            "408",
            "409",
            "410",
            "411",
            "412",
            "413",
            "414",
            "415",
            "416",
            "417",
            "418",
            "419",
            "420",
            "421",
            "422",
            "423",
            "424",
            "426",
            "428",
            "429",
            "431",
            "451",
            "500",
            "501",
            "502",
            "503",
            "504",
            "505",
            "506",
            "507",
            "508",
            "509",
            "510",
            "511"
          ]
        },
        "statusCodeValue": {
          "type": "integer",
          "format": "int32"
        }
      }
    }
  }
}